requirements.txt
pony==0.7.13
fastapi==0.54.1
python-dotenv==0.13.0
psycopg2-binary==2.8.5
email_validator==1.1.0
pyjwt==1.7.1
passlib==1.7.2
bcrypt==3.1.7